Fix: chat endpoint fixes, webockets endpoint connection fix

+++ backend/services/redis_service/service.py
@@ -303,13 +303,14 @@ class RedisService:
         return [await self.__client.get(key) async for key in self.__client.scan_iter(match=f"{first_prefix}{user_id}{self.__exclude_chat_prefix_2}*")]
 
 
-    @redis_error_handler
+    # @redis_error_handler
     async def clear_exclude_chat_ids(self, user_id: str, exclude_type: ChatType) -> None:
         # Caution! May not be optimized for large keys numbers
         first_prefix = self._get_right_first_exclude_chat_prefix(exclude_type=exclude_type)
         to_delete = [key async for key in self.__client.scan_iter(match=f"{first_prefix}{user_id}{self.__exclude_chat_prefix_2}*")]
 
-        self.__client.delete(*to_delete)    
+        if to_delete:
+            await self.__client.delete(*to_delete)    
 
     @redis_error_handler
     async def save_chat_token(self, chat_token, user_id: str) -> None:
